update: character repo, service

+++ lepl/src/main/java/com/lepl/domain/character/Exp.java
@@ -17,5 +17,5 @@ public class Exp {
     private Long id;
 
     private Long expAll;
-    private int level;
+    private Long level;
 }
